refactor: 정적 파일 response 반환 책임을 ResourceProvider 에게 위임, 컨트롤러 response 반환 책임을 HandlerMapper 에게 위임
